feat: switch to using shipped assets
