+++ python/pyproject.toml
@@ -2,11 +2,7 @@
 name = "squishy-volumes-extension"
 version = "0.0.0" # not used
 requires-python = ">=3.13,<3.14"
-dependencies = [
-  "vendorize",
-  "pip>=25.3",
-  "tree-clipper==0.1.12",
-]
+dependencies = []
 classifiers = ["Private :: Do Not Upload"]
 
 [project.optional-dependencies]
@@ -16,12 +12,5 @@ dev = [
   "requests==2.32.5",
   "fake-bpy-module>=20251003",
   "numpy==1.26.4", # should match what blender uses
+  "tree-clipper==0.1.12",
 ]
-
-[tool.uv.sources]
-# to get support for pyproject.toml (which isn't released)
-vendorize = { git = "https://github.com/mwilliamson/python-vendorize.git", rev = "04d1ae6c636666ea78447134c37e41c082f15190" }
-
-[tool.vendorize]
-target = "src/squishy_volumes_extension/_vendor"
-packages = ["tree-clipper"]
